refactor(bookstore): split BookBuySection into server/client components
BookBuySection is now a server component (Buy header + OOS banner);
interactive format selector extracted to BookBuyFormats client component.
Reduces client JS bundle by moving static markup to server rendering.
